changed to float

+++ gci.c
@@ -2,7 +2,7 @@
 
 int main() {
 int a ,c;
-double b=1;
+float b=1;
 printf("enter the number whose factorial is to be found");
 scanf("%d",&a);
 for(int n=1;n<=a;n++){
@@ -10,5 +10,5 @@ b=b*n;
 
 }
 printf("factorial of given number is ");
-printf("%.0lf",b);
+printf("%.0f",b);
 }
